AUTHORIZATION - authorize user to edit or delete questions using gate method

+++ app/Http/Controllers/QuestionController.php
@@ -64,7 +64,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        return  view('questions.edit',compact('question'));
+        if(\Gate::allows('update-question',$question)) {
+            return view('questions.edit', compact('question'));
+        }
+        abort(403,'Access Denied.');
     }
 
     /**
@@ -88,7 +91,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
-      $question->delete();
-      return back()->with('success','Your question has been deleted.');
+        if(\Gate::allows('delete-question',$question)) {
+            $question->delete();
+            return back()->with('success','Your question has been deleted.');
+        }
+        abort(403,'Access Denied.');
     }
 }
